Update onedayfullcolaug.py
For an entire day, this code collocates all the INSAT-3DR 1B-IMAGER and CloudSat 2B-CLDCLASS files. Expected execution time = 40-100 hours, producing 10-13 collocated files.

By simply adding a comparison for date as well, this can work for even folders which have data for entire months or entire years, however it's better to add parallelization in that case for the time it'd take in that case. 

swir radiance is not included in this data. It can be added with few additional lines of code, the steps are as follows:-
1- create swir radiance array.
2- create an empty swirradcollocated list, example at line no. 183.
3 - Find the swirrad collocated value using indexvis.
4 - Append it to swirradcollocated list.
5 - Add it to the dataframe.